moved 2 doppler factor associated tests to the test_doppler_factor
