+++ tardis/transport/tests/test_doppler_factor.py
@@ -2,6 +2,7 @@ import pytest
 import numpy as np
 
 import tardis.transport.frame_transformations as frame_transformations
+import tardis.montecarlo.montecarlo_numba.r_packet as r_packet
 
 from numpy.testing import (
     assert_almost_equal,
@@ -112,3 +113,29 @@ def test_get_inverse_doppler_factor_full_relativity(mu, beta, expected):
         mu, beta
     )
     assert_almost_equal(obtained, expected)
+
+
+@pytest.mark.parametrize(["mu", "r", "inv_t_exp"], [(-0.3, 5, 1e10)])
+def test_unphysical_doppler_factor(mu, r, inv_t_exp):
+    # Set the params from test cases here
+    # TODO: add relativity tests
+    time_explosion = 1 / inv_t_exp
+
+    # Perform any other setups just before this, they can be additional calls
+    # to other methods or introduction of some temporary variables
+    with pytest.raises(r_packet.SuperluminalError):
+        obtained = frame_transformations.get_doppler_factor(
+            r, mu, time_explosion
+        )
+
+
+@pytest.mark.parametrize(["mu", "r", "inv_t_exp"], [(-0.3, 5, 1e10)])
+def test_unphysical_inverse_doppler_factor(mu, r, inv_t_exp):
+    # Set the params from test cases here
+    # TODO: add relativity tests
+    time_explosion = 1 / inv_t_exp
+
+    # Perform any other setups just before this, they can be additional calls
+    # to other methods or introduction of some temporary variables
+    with pytest.raises(r_packet.SuperluminalError):
+        obtained = r_packet.get_inverse_doppler_factor(r, mu, time_explosion)
